Code is shorter but harder to read, success!
But i successfully removed the second instance of '.py'!

I feelt it should be unecessary to say the same thing twice.
To check that it has '.py' and then remove it.

So this version does both in one but is harder to read for no real
benefit so not production code..